Updated pre-deploy script
predeploy script should not fail if one of the three files are absent.

+++ predeploy.py
@@ -4,6 +4,8 @@ import subprocess
 
 files = ["requirements.txt", "Procfile", "DOKKU_SCALE"]
 for f in files:
-    shutil.copyfile(os.path.join("apps", os.environ["DASH_APP_NAME"], f), f)
+    app_file_path = os.path.join("apps", os.environ["DASH_APP_NAME"], f)
+    if os.path.exists(app_file_path):
+        shutil.copyfile(os.path.join("apps", os.environ["DASH_APP_NAME"], f), f)
 
 subprocess.run("python -m pip install -r requirements.txt".split(" "))
